+++ scripts/build-js.js
@@ -4,19 +4,32 @@ const fs = require('fs');
 const path = require('path');
 const mkdirp = require('mkdirp');
 const pify = require('pify');
-const globby = require('globby');
 const UglifyJS = require('uglify-js');
 const optimizeJs = require('optimize-js');
 const timelog = require('./timelog');
 const buildSvgLoader = require('./build-svg-loader');
 
-const jsGlob = path.join(__dirname, '../src/js/*.js');
+const jsSrcDir = path.join(__dirname, '../src/js');
 
-function buildJs(outfile, options) {
+// If you create a new script for assembly.js you must add it here
+const jsFiles = [
+  path.join(jsSrcDir, 'focus-control.js'),
+  path.join(jsSrcDir, 'icon-functions.js'),
+];
+
+/**
+ * Build JS.
+ *
+ * @param {Object} [options]
+ * @param {string} [options.outfile] - Path to which built JS should be written.
+ * @param {Object} [options.quiet] - Suppress logs.
+ * @return {Promise<void>}
+ */
+function buildJs(options) {
   options = options || {};
 
   if (!options.quiet) timelog('Building JS');
-  outfile = outfile || path.join(__dirname, '../dist/assembly.js');
+  const outfile = options.outfile || path.join(__dirname, '../dist/assembly.js');
 
   return Promise.all([
     buildSvgLoader(),
@@ -24,13 +37,12 @@ function buildJs(outfile, options) {
   ])
     .then((data) => {
       const allJs = data.join('');
-      return UglifyJS.minify(allJs, { fromString: true }).code;
-    })
-    .then((minifiedJs) => {
+      if (options.unminified) return allJs;
+      const minifiedJs = UglifyJS.minify(allJs, { fromString: true }).code;
       return optimizeJs(minifiedJs);
     })
     .then((optimizedJs) => {
-      return pify(mkdirp)(path.join(__dirname, '../dist')).then(() => {
+      return pify(mkdirp)(path.dirname(outfile)).then(() => {
         return pify(fs.writeFile)(outfile, optimizedJs);
       });
     }).then(() => {
@@ -39,14 +51,13 @@ function buildJs(outfile, options) {
 }
 
 function concatJs() {
-  let result = '';
-  return globby(jsGlob).then((jsFiles) => {
-    return Promise.all(jsFiles.map((jsFile) => {
-      return pify(fs.readFile)(jsFile, 'utf8').then((jsFileContent) => {
-        result += jsFileContent;
-      });
-    }));
-  }).then(() => result);
+  const result = [];
+  // Deterministic order needed for tests
+  return Promise.all(jsFiles.map((jsFile, index) => {
+    return pify(fs.readFile)(jsFile, 'utf8').then((jsFileContent) => {
+      result[index] = jsFileContent;
+    });
+  })).then(() => result.join(''));
 }
 
 module.exports = buildJs;